The module and driver routines now work correctly for the HHWind and FFWind files. The driver currently does not do much since some features are not fully implimented
There are some issues regarding how the Types files are generated. Right now, the InflowWind_Types.f90 file generates mostly ok if the Reg-IfW_FFWind_list.txt (modified version of Reg-IfW_FFWind.txt for include) and its counterpart for HHWind are used. After that, a pair of lines for USE IfW_FFWind_Types must be added to the start of the InflowWind_Types.f90 file for it to work ok.

More instructions on how to actually integrate this module into the glue code will be added.


git-svn-id: https://windsvn2.nrel.gov/InflowWind/branches/modularization@64 96535b8f-fecd-4773-a067-ac893715d582