Added the code for handling the OutList given in the input file.  The driver doesn't yet support this, but the code appears to be working.
The OutList and WriteOutput code is perhaps a bit suboptimal, but I think it is functional.

Also made a few changes in InflowWind regarding how the coordinate transforms are handled.  The matrices for the rotation are now handled in the Init routine and stored in OtherStates.  The WindViXYZ coordinates are copied in OtherStates to WindViXYZprime with the applied coordinate transform during Initialization.  This simplifies the CalcOutput routine slightly.  NOTE: This has not been thouroughly tested yet!!!!



git-svn-id: https://windsvn2.nrel.gov/InflowWind/branches/modularization2@145 96535b8f-fecd-4773-a067-ac893715d582